Add a Candidate struct with a .seed function
A ‘candidate’ is a partial superpermutation that
will be contained in the tree. It does not contain
enough information to reconstruct the string, but
just enough to proceed with the algorithm.

Each candidate tracks which permutations have been
seen so far in a BitSet. The Lehmer crate allows
permutations to be represented as decimal numbers,
so we can use this to compactly store this
information.

We store the tail of the string which will allow
us to build the next candidates and how many
symbols have been ‘wasted’ in the string. A symbol
Is wasted if it does not add a new permutation.