Implement Candidate#expand_one
A candidate string can be expanded to build the
next candidate strings with different symbols.

We can short-circuit this step if the tail of the
current or next candidate is not n-1 symbols long
because this means new permutations definitely
haven’t been added to the string.

If these checks succeed, we construct the
permutation, convert it to a decimal and check if
we’ve seen it before in the BitSet.