Fix a bug with detecting wasted symbols
If the tail is [1, 2, 3, 4] and the next symbol is
a 1 then the new tail will be [2, 3, 4, 1]. It is
not sufficient to only check if the new tail is
shorter than n-1 symbols. We also need to check
the first symbol of the previous tail is not equal
to the last symbol of the new tail.