Implement Candidate#expand
Adds a function that expands the next set of
candidates for an existing candidate. There’s no
need to expand the candidate that ends with same
symbol as any superpermutation that contains two
of the same symbol in a row can be made smaller by
removing one of those symbols.