Add a Frontier struct
The plan is to implement a best-first search of
candidates, prioritised by their total waste and
the number of permutations they contain.

Once a candidate is expanded, it is no longer
needed and can be removed from the search tree.
That means the candidates form the leaves of a
tree, which is sometimes called the ‘frontier’ of
the search.