Add Frontier#prune
We need to be able to prune Candidates from the
Frontier based on the threshold from the Bounds.
If a Candidate for a given number of wasted
symbols has fewer permutations than the threshold
we should prune it.

We have a couple of options in how we do this:

1. We can prune just the candidates for the number
   of wasted symbols currently being expanded

2. We can eagerly prune the candidates for higher
   numbers of wasted symbols

In theory, option 1. should be faster because the
candidates with higher waste should have no
bearing on the current part of the search. When
the search moves on to the next number of wasted
symbols, it can prune those candidates then. So
there’s no point repeatedly pruning these
candidates now.

However, option 2. might be better if it’s costly
to keep these candidates in memory, e.g. if we’re
running out of memory and the operating system
starts swapping to disk.