Add Bounds#found_for_superpermutation
When we’ve found the lower bound for the number of
wasted symbols that can fit all N! permutations,
we’ve found the shortest possible superpermutation
for N. We’ll use this function to determine when
to stop the algorithm.

We may as well set the the upper bound to N!
rather than usize::MAX as this is a tighter bound.