Don’t actually remove candidates when pruning
As the algorithm runs, we prune candidates that
cannot improve on the highest permutations for the
current number of wasted symbols.

However, the choice of which candidates is pruned
is based on the current lower and upper bounds for
the current number of wasted symbols.

As the search moves on to higher numbers of wasted
symbols, their bounds can mean it’s worth
exploring these candidates again. We can’t do that
if we’ve completely removed them from the search.

Therefore, this commit makes it so pruning
actually just moves candidates from one queue to
another so they can potentially be ‘unpruned’
later.