Unprune candidates that can improve the lower bound
Just before the Frontier moves on to finding the
bounds for the next wasted symbol, we unprune
candidates that could potentially improve the
lower bound.

The way these candidates are determined is a bit
complicated. Here’s how it works:

1) Get the lower and upper bounds for the current
number of wasted symbols, e.g. 4 waste: 14..=16

2) For each smaller number of wasted symbols,
calculate how many additional wasted symbols we
could add, e.g. for waste 3, we could add 1 more

3) Look up the maximum number of permutations 1
wasted symbol could yield, e.g. 8 permutations

4) Subtract this number from the current bounds
and offset by 1, e.g. (14..=16) -> (7..=8)

So for candidates that waste 3 symbols, we should
re-introduce those with 7 and 8 permutations as
they have the potential to increase the lower
bound for the current number of wasted symbols.

The reason we offset by 1 is because we’re trying
to improve the lower bound. If we didn’t do this,
we’d re-add candidates that can only equal the
current best candidates, which doesn’t help.

It makes sense to incrementally unprune buckets
one-by-one, rather than all at once. If we get
lucky and a candidate increases the lower bound,
this reduces (or eliminates) the remainder of what
needs to be unpruned.

This leaves a dilemma of what is the best order to
unprune candidates. Either we choose to favour
the deepest candidates in the Frontier, or those
that are at the high-end of the bounds for
permutations at each number of wasted symbols.

After some quick benchmarking, it seems like the
latter approach is marginally faster, but this
needs more investigation.