Split the writing of a single bucket into multiple files
If data is written for the same bucket, it now
goes into files suffixes by an index so that we
don’t overrite what’s currently there. That means
we need to glob the directory to find the name of
the file to read from.

Example filename:

candidates-with-3-wasted-symbols-and-4-permutations.dat.gz.0