Swap buckets to and from disk when memory is low
The algorithm is very memory thirsty as it holds a
lot of candidates in its disabled queue. Most of
these aren’t actively being used and can be
offloaded to disk.

Unfortunately, this is really slow and for the
majority of the time the algorithm is running,
it’s offloading memory, which is a shame. I’ve
tried to make this faster by using Rayon to
parallelise it, with limited success.

Without gzipping the offloaded memory, the
algorithm consumes about 70GB of disk space per
hour. Space (either in memory or on disk) is
definitely the limiting factor for running this
algorithm for an extended period of time.