Write to disk in chunks for large buckets
Later on in the search, buckets start to get very
large so limit the maximum file size we write to
disk so that we can unprune bits of that bucket
incrementally.

In theory, this should reduce churn on reading in
a huge file- only to have to write it back out
again when memory is exceeded.

222_222 is arbitrary and limits file sizes to
about 50MB for N=6.